Importing Linux 2.6.22
I'm using this version for now to get a known good base to try and work
from.  Once I get this working on the FPGA, I hope to fast forward to a
more modern kernel and see if I can't get those working.  If that works,
I'll be able to develop against a modern version instead of a super old
one, which will likely simplify the number of trees I have to maintain
for different ports of LockBox on different architectures.