Import and forward port sunhv_net into tree.
This driver originally came from the OpenSparc T1 tree, where it lives @
design/sys/edk/os/Ubuntu/7.10-Gutsy/src/drivers/sunhv_net for reasons
known only to the engineers who decided that was the appropriate place
to put it.

In any case, this commit takes that driver and puts it into my kernel
tree where it belongs.  Quite a bit has changed in the five years since
this driver was written for 2.6.22 and forward porting it to 3.9.0
included some interface changes.  This driver mostly copies the sunqe
driver since it seemed to be the one that was closest.

It not only compiles, but even appears to be capable of working and
sending packets, though often after some fiddling.  The device does not
appear to come up with a MAC address, but setting one (ifconfig eth[0-9]
hw ether <mac>) and then bringing the interface up seems to work.  This
revision even appears to correctly report the number of packets the
interface transmits and receives.

I am sure bugs remain, but this driver is sufficient to send packets and
given that I haven't slept since yesterday, that seems worth not only
celebrating, but committing.