Implement frequency hack: time can now take normal amount of time
Implements the CONFIG_OPENSPARC_T1_FPGA_FREQ_HACK option, which enables
a kernel hack that divides the frequency the OpenSPARC T1 core reports
by 20.  This matches the default FPGA implementation of the OpenSPARC T1
core, which reports a 1GHz clock frequency even though it runs at 50MHz.

Since the system assumes that a 1GHz processor works at 1GHz, all the
timers on the system take 20x longer, which can be extremely annoying.
So this hack makes time take a normal amount of time on this platform.

If you enable this option on any other SPARC system time will run 20x
faster than it should.  So don't do that.

If you change the clock frequency of your FPGA, time will be askew
again.  This is just a hardcoded hack to change the frequency to the
default FPGA implementation's target clock speed.